back-end/requirements.txt
flask
flask-restful
requests
flask-mongoengine
flask-cors
